Use shlex to split args to handle paths with spaces
Additionally quote the strings used in building the paths.

Without this change paths or files containing spaces would be split as
separate arguments, which would result in a failure to the call of
PyQt5.pyrcc_main e.g.:

    PyQt5/pyrcc_main.py: File does not exist "/Users/xxx/path_before_space"

Where the rest of the path after the space is treated as the next
argument and hence results in a 'File does not exist' error.